+++ src/items/components/enterable.rs
@@ -1,0 +1,6 @@
+use bevy::prelude::Component;
+
+/// Component we can "enter" on
+/// Will trigger an OnEnter event
+#[derive(Component)]
+pub struct EnterAble;
