+++ src/items/components/ghost_only.rs
@@ -2,6 +2,3 @@ use bevy::prelude::*;
 
 #[derive(Component)]
 pub struct GhostOnly;
-
-#[derive(Component)]
-pub struct PlayerOnly;
