+++ src/items/components/is_usable.rs
@@ -1,4 +1,4 @@
-use bevy::prelude::*;
+use bevy::prelude::Component;
 
 #[derive(Component)]
 pub struct IsUsable;
