+++ src/items/components/level_teleporter.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::Component;
+
+#[derive(Component)]
+pub struct LevelTeleporter(pub String);
