+++ src/items/components/player_only.rs
@@ -2,6 +2,3 @@ use bevy::prelude::*;
 
 #[derive(Component)]
 pub struct PlayerOnly;
-
-#[derive(Component)]
-pub struct SingleUse;
