+++ src/items/mod.rs
@@ -1,16 +1,23 @@
-pub mod debug_name;
-pub mod dependencies;
-pub mod enterable;
 pub mod events;
-pub mod ghost_only;
-pub mod is_usable;
-pub mod level_teleporter;
-pub mod player_only;
 pub mod populate_items;
-
 pub mod systems {
-    pub mod is_activated;
-    pub mod people_on;
-    pub mod teleport;
+    pub mod count_people_on_system;
+    pub mod level_teleporter_system;
+    pub mod people_enter_system;
+    pub mod teleporter_system;
     pub mod toggle;
+    pub mod update_is_activated_system;
+}
+pub mod components {
+    pub mod debug_name;
+    pub mod dependencies;
+    pub mod enterable;
+    pub mod ghost_only;
+    pub mod is_activated;
+    pub mod is_usable;
+    pub mod level_teleporter;
+    pub mod people_on;
+    pub mod player_only;
+    pub mod single_use;
+    pub mod teleporter;
 }
