+++ src/items/populate_items.rs
@@ -1,14 +1,15 @@
 use crate::constantes::CELL_LENGTH;
-use crate::items::debug_name::DebugName;
-use crate::items::dependencies::Dependencies;
-use crate::items::enterable::EnterAble;
-use crate::items::ghost_only::GhostOnly;
-use crate::items::is_usable::IsUsable;
-use crate::items::level_teleporter::LevelTeleporter;
-use crate::items::player_only::{PlayerOnly, SingleUse};
-use crate::items::systems::is_activated::IsActivated;
-use crate::items::systems::people_on::PeopleOn;
-use crate::items::systems::teleport::Teleporter;
+use crate::items::components::debug_name::DebugName;
+use crate::items::components::dependencies::Dependencies;
+use crate::items::components::enterable::EnterAble;
+use crate::items::components::ghost_only::GhostOnly;
+use crate::items::components::is_activated::IsActivated;
+use crate::items::components::is_usable::IsUsable;
+use crate::items::components::level_teleporter::LevelTeleporter;
+use crate::items::components::people_on::PeopleOn;
+use crate::items::components::player_only::PlayerOnly;
+use crate::items::components::single_use::SingleUse;
+use crate::items::components::teleporter::Teleporter;
 use crate::items::systems::toggle::ToggleInteract;
 use crate::map_parser::map_repr::{InteractionType, ObjectRepr, ObjectType};
 use crate::math::vec2i::Vec2i;
