+++ src/items/systems/level_teleporter_system.rs
@@ -1,13 +1,11 @@
-use crate::items::enterable::EnterAble;
+use crate::items::components::enterable::EnterAble;
+use crate::items::components::is_usable::IsUsable;
+use crate::items::components::level_teleporter::LevelTeleporter;
+use crate::items::components::player_only::PlayerOnly;
 use crate::items::events::OnEnterEvent;
-use crate::items::is_usable::IsUsable;
-use crate::items::player_only::PlayerOnly;
 use crate::level::components::level_to_go::LevelToGo;
 use crate::state::GameState;
 use bevy::prelude::*;
-
-#[derive(Component)]
-pub struct LevelTeleporter(pub String);
 
 pub fn level_teleporter_system(
     mut on_enter_event_reader: EventReader<OnEnterEvent>,
