+++ src/level/plugin.rs
@@ -1,12 +1,12 @@
-use crate::items::enterable::people_enter_system;
+use crate::items::components::ghost_only::GhostOnly;
+use crate::items::components::player_only::PlayerOnly;
 use crate::items::events::OnEnterEvent;
-use crate::items::ghost_only::GhostOnly;
-use crate::items::level_teleporter::level_teleporter_system;
-use crate::items::player_only::PlayerOnly;
-use crate::items::systems::is_activated::update_is_activated_system;
-use crate::items::systems::people_on::count_people_on_system;
-use crate::items::systems::teleport::{teleporter_activate_system, teleporter_system};
+use crate::items::systems::count_people_on_system::count_people_on_system;
+use crate::items::systems::level_teleporter_system::level_teleporter_system;
+use crate::items::systems::people_enter_system::people_enter_system;
+use crate::items::systems::teleporter_system::{teleporter_activate_system, teleporter_system};
 use crate::items::systems::toggle::{toggle_on_enter_system, toggle_on_interact_system};
+use crate::items::systems::update_is_activated_system::update_is_activated_system;
 use crate::level::load_level::load_level;
 use crate::level::unload_level::unload_level;
 use crate::player::interact::{GhostInteractEvent, PlayerInteractEvent};
